app/auth/forms.py
#app/auth/forms.py
from flask_wtf import FlaskForm
from wtforms import StringField, PasswordField, BooleanField, SubmitField
from wtforms.validators import DataRequired, Email , EqualTo, ValidationError, Length
from app.models import User